+++ plugin/controllers/defaults.py
@@ -141,6 +141,10 @@ EXT_EVENT_INFO_SOURCE = getExtEventInfoProvider()
 
 TRANSCODING = getTranscoding()
 
+VXGENABLED = isfile(getPublicPath("/vxg/media_player.pexe"))
+
+WEBTV = VXGENABLED or TRANSCODING
+
 
 def getOpenwebifPackageVersion():
 	control = glob('/var/lib/opkg/info/*openwebif.control')