+++ CHANGES.md
@@ -2,6 +2,7 @@
 * move about to left menu
 * fix ldc screenshot
 * fix screenhot javascript error
+* show webtv only if transcoding and vgx is available
 
 ## Version 2.0.0
 * remove all remote control images